Add class to handle form submit
Task: AFORM-3697
