+++ .vscode/settings.json
@@ -1,4 +1,4 @@
 {
-    "jestrunner.jestCommand": "npm run test --",
     "jest.rootPath": "src/@optimizely/forms-sdk",
+    "jest.jestCommandLine": "npm run test --",
 }
