--- src/@optimizely/forms-react/src/models/index.ts
@@ -1,1 +1,0 @@
-export * from "./SubmitButtonType";
