+++ src/@optimizely/forms-sdk/src/index.ts
@@ -4,3 +4,5 @@ export * from "./form-step";
 export * from "./helpers";
 export * from "./form-validator";
 export * from "./form-depend-conditions";
+export * from "./form-storage";
+export * from "./form-submit";
