+++ src/@optimizely/forms-sdk/src/models/enums/index.ts
@@ -1,4 +1,5 @@
 export * from "./ValidatorType";
 export * from "./ConditionCombinationType";
 export * from "./SatisfiedActionType";
-export * from "./ConditionFunctionType"
+export * from "./ConditionFunctionType";
+export * from "./SubmitButtonType";
