--- src/@optimizely/forms-sdk/src/models/states/FormDependencies.ts
@@ -1,4 +1,0 @@
-export interface FormDependencies {
-    elementKey: string
-    isSatisfied: boolean
-}
