+++ src/@optimizely/forms-sdk/src/models/states/FormState.ts
@@ -1,5 +1,4 @@
 import { FormContainer } from "../FormContainer"
-import { FormDependencies } from "./FormDependencies"
 import { FormSubmission } from "./FormSubmission"
 import { FormValidation } from "./FormValidation"
 import { StepDependencies } from "./StepDependencies"
@@ -8,7 +7,7 @@ export interface FormState {
     isReset: boolean
     formSubmissions: FormSubmission[]
     formValidations: FormValidation[]
-    formDependencies: FormDependencies[]
     stepDependencies: StepDependencies[]
     formContainer: FormContainer
+    dependencyInactiveElements: string[]
 }
