+++ src/@optimizely/forms-sdk/src/models/states/index.ts
@@ -1,5 +1,4 @@
 export * from "./FormState";
-export * from "./FormDependencies";
 export * from "./FormSubmission";
 export * from "./FormValidation";
 export * from "./StepDependencies";
